+++ CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes will be documented in this file.
 ### Added
 
 * `NotSupported` exception (clean-up)
-* `CustomMetrics` (#8)
+* `CustomMetrics` (publish) and `CustomMetricsUpdate` (receive) (#8)
 * `NotImplemented` exception (clean-up)
 * `Error::CONDITIONAL_REQUEST_HAS_FAILED` and `Error::UNKNOWN_ORDER_ID` (#25)
 * `utils::was_order_received` and `utils::to_request_status` (#25)